Add files via upload
I updated the priors after looking at the MCM results to increase the acceptance rate. 
I modified the check of the critical points adding the condition that with 4 equilibria the saddles need to be ordered in a specific way. For that, we need to pass the parameters to the check function. I called v2 the ones with these midifications.